HTBHF-1065 Moved app started event logging to ClaimantServiceApplication
