+++ api/src/main/java/uk/gov/dhsc/htbhf/claimant/ApiDocumentation.java
@@ -1,13 +1,10 @@
 package uk.gov.dhsc.htbhf.claimant;
 
 import lombok.extern.slf4j.Slf4j;
-import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.beans.factory.annotation.Value;
-import org.springframework.boot.context.event.ApplicationReadyEvent;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.context.annotation.Import;
-import org.springframework.context.event.EventListener;
 import springfox.bean.validators.configuration.BeanValidatorPluginsConfiguration;
 import springfox.documentation.builders.PathSelectors;
 import springfox.documentation.builders.RequestHandlerSelectors;
@@ -15,8 +12,6 @@ import springfox.documentation.service.ApiInfo;
 import springfox.documentation.service.Contact;
 import springfox.documentation.spi.DocumentationType;
 import springfox.documentation.spring.web.plugins.Docket;
-import uk.gov.dhsc.htbhf.logging.EventLogger;
-import uk.gov.dhsc.htbhf.logging.event.ApplicationStartedEvent;
 
 import static java.util.Collections.emptyList;
 
@@ -31,15 +26,6 @@ public class ApiDocumentation {
 
     @Value("${app.version:}") // use APP_VERSION env variable if available, otherwise give no version info
     private String appVersion;
-
-    @Value("${instance.index:}") // use INSTANCE_INDEX env variable if available, otherwise give no index info
-    private String instanceIndex;
-
-    @Value("${vcap.application.application_id:}") // the id of the application as provided by cf
-    private String applicationId;
-
-    @Autowired
-    private EventLogger eventLogger;
 
     @Bean
     public Docket apiDocket() {
@@ -64,15 +50,5 @@ public class ApiDocumentation {
                 emptyList()
         );
 
-    }
-
-    @EventListener(ApplicationReadyEvent.class)
-    public void logAfterStartup() {
-        eventLogger.logEvent(ApplicationStartedEvent.builder()
-                .applicationId(applicationId)
-                .applicationVersion(appVersion)
-                .instanceIndex(instanceIndex)
-                .build()
-        );
     }
 }
